Module Slurp__Route.Path

Path operators.

val path : string -> ('a, 'b, 'c) path -> ('a, 'b, 'c) path

path s starts a path with the given root s. To defines an empty root prefix simply use "".

Be warned that s must be an atomic file or directory name (ie. without a directory separator like '/' under Unix based systems or '\' under Windows). Use cut in this case.

val cut : string -> ('a, 'b, 'c) path -> ('a, 'b, 'c) path

cut s splits automatically the given path to its underlying segments to form a valid path.

val (/) : (('a, 'b, 'c) path -> ('d, 'b, 'c) path) -> string -> ('a, 'b, 'c) path -> ('d, 'b, 'c) path

a / b matches the path a then b.

val (/*/) : (('a, 'b, 'c) path -> ('d, 'b, 'c) path) -> string -> ('a, 'b, 'c) path -> ('d, 'b, 'c) path

a /*/ b matches the path a then any section and then b.

val (/**/) : (('a, 'b, 'c) path -> ('d, 'b, 'c) path) -> string -> ('a, 'b, 'c) path -> ('d, 'b, 'c) path

a /**/ b matches the path a then any sections (if any) and then b.

val (/:) : (('a -> 'b, 'c, 'd) path -> ('e, 'c, 'd) path) -> 'a parameter -> ('b, 'c, 'd) path -> ('e, 'c, 'd) path

a /: b matches the path a then the bind the b to its future instanciation.

val (/*/:) : (('a -> 'b, 'c, 'd) path -> ('e, 'c, 'd) path) -> 'a parameter -> ('b, 'c, 'd) path -> ('e, 'c, 'd) path

a /*/: b matches the path a, then any section and then bind the b to its future instanciation.

val (/**/:) : (('a -> 'b, 'c, 'd) path -> ('e, 'c, 'd) path) -> 'a parameter -> ('b, 'c, 'd) path -> ('e, 'c, 'd) path

a /**/: b matches the path a, then any sections (if any) and then bind the b to its future instanciation.

val (/?) : (('a, 'b, 'c) path -> ('d, 'b, 'c) path) -> 'e -> (('a, 'b, 'c) query -> ('d, 'b, 'c) path) * 'e

a /? b matches the path a then the query part b.

val (/*/?) : (('a, 'b, 'c) path -> ('d, 'b, 'c) path) -> 'e -> (('a, 'b, 'c) query -> ('d, 'b, 'c) path) * 'e

a /*/? b matches the path a, then any section and then the query part b.

val (/**/?) : (('a, 'b, 'c) path -> ('d, 'b, 'c) path) -> 'e -> (('a, 'b, 'c) query -> ('d, 'b, 'c) path) * 'e

a /**/? b matches the path a, then any sections (if any) and then the query part b.

val (+) : ((('b -> 'c, 'd, 'e) query -> 'f) * 'b parameter) -> 'a parameter -> (('c, 'd, 'e) query -> 'f) * 'a parameter

a + b adds a to query parameters of b.

val (-->) : ((('b -> 'c, 'b, 'c) query -> ('d, 'b, 'c) path) * 'b body) -> 'c response -> ('d, 'b, 'c) path

a --> b finalize the path specification and adds a as the body specification and b as the response specification.